third_party/blink/web_tests/http/tests/inspector-protocol/issues/third-party-cookie-blocking-first-party-set-expected.txt

Test third-party cookie blocking between sites in same First-Party Set
Test started

Running test: thirdPartyCookieBlockingDisabled
Number of cookies: 0
Number of cookies: 1
Cookie not blocked: true

Running test: thirdPartyCookieBlockingEnabled
Number of cookies: 0
Number of cookies: 1
Cookie blocked: true
Reasons: UserPreferences
